all_requirements.txt
amqp==2.6.1
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
boto3==1.17.17
botocore==1.20.112
celery==4.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
coverage==6.0.2
cryptography==35.0.0
defusedxml==0.7.1
Django==3.1.7
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.1.0
django-haystack==3.1.1
django-js-asset==1.2.2
django-machina==1.1.5
django-mptt==0.13.4
django-redis-cache==2.1.0
django-rq==2.1.0
django-storages==1.11.1
django-taggit==1.2.0
django-timezone-field==4.2.1
django-translation-manager==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.2
idna==3.3
iniconfig==1.1.1
invoke==0.21.0
jmespath==0.10.0
kombu==4.6.11
Markdown==2.6.9
markdown2==2.4.1
more-itertools==8.4.0
#mysqlclient==1.4.6
nose==1.3.7
notifications-plugin==0.1.2
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pluggy==1.0.0
polib==1.1.1
psutil==5.8.0
py==1.10.0
pycparser==2.20
Pygments==2.2.0
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.5
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
qrcode==7.3.1
redis
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.10.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
selenium==2.53.6
six==1.16.0
social-auth-app-django==3.1.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
tornado==4.5.3
urllib3==1.26.7
vine==1.3.0
